microcloud/service/lxd: Use lxd.InstanceServer for LXD client
